# RingCT cache (huge block propagation speedup)