The problem: block propagation takes up to 1.5 seconds even on a fast 6-core Ryzen with 64 GB RAM and an NVMe SSD because monerod must check all transactions before adding them to the new block. This affects mining pools (including P2Pool) and solo miners the most by increasing block orphan rate.

I discovered it when trying to optimize my P2Pool nodes. What I found is each transaction can be checked by monerod up to 3 times:

- when it's added to the mempool
- when it's added to the new block
- when it's sent via ZMQ to connected P2Pool nodes

Each check does exactly the same calculations, so it can be cached. The main offender here is `verRctNonSemanticsSimple` which doesn't have side effects, so if it gets the same input more than one time, it will return the same result each time. This PR adds a cache of last 8192 `rctSig` values that passed the check. It doesn't store the values themselves, but their keccak (`cryptonote::get_blob_hash`) hashes to save memory. Using keccak should be safe from collision attacks since it's a cryptographically secure hash.

I've been running this code on my nodes for a few days and here is a comparison between 2 nodes with the same hardware (6 core CPU, 64 GB RAM, NVMe SSD):
|Block|tx count|propagation delay|optimized delay|speedup
|-|-|-|-|-
|2772901|117|0.6696 s|0.1059 s|6.3x
|2772903|45|0.2905 s|0.059 s|4.9x
|2772907|75|0.5127 s|0.0926 s|5.5x
|2773001|3|0.0766 s|0.0406 s|1.9x
|2773021|112|1.2886 s|0.1713 s|7.5x
|2773043|159|0.8643 s|0.1359 s|6.3x

Raw data: https://paste.debian.net/hidden/cefbf13b/

@UkoeHB and @kayabaNerve - one of you guys should review this.