@UkoeHB 2022-12-12
This PR breaks `performance_tests/check_tx_signature.h`. I'd much rather see the cache be a member variable owned by the `Blockchain` class. Then this line would change:
```
-      if (!rct::verRctNonSemanticsSimple(rv))
+      if (!m_rct_proof_validation_result_cache.verify(rv))
```