@SChernykh 2022-12-12
But I want tests to run this code, this is how I found serialization issues (see conversation above). Maybe it's better to have something like `rct::verRctClearCache()` and call it where necessary?