@kayabaNerve 2022-12-12
I would also prefer explicit tests regarding copying RCT sigs to different TXs/TXs only different by the key images/etc. Even though they may not be comprehensive, they'd demonstrate thought and establish a baseline.

We could also write a test which checks any edit to any field of a TX (prefix, signatures, RCT base, RCT prunable...) changes its sig hash and is accordingly included in the above message serialization.

This is a great efficiency improvement yet it does so by cutting out a lot of the checks (which is fine when proper). Hence why I want to be very sure it is proper.