@UkoeHB 2022-12-12
> But I want tests to run this code, this is how I found serialization issues (see conversation above).

In that case, you should write an actual test for this cache. If the `Blockchain` thing is not properly tested, then add tests for it.