@SChernykh 2022-12-12
> Users of the ringct library should not be forced to depend on/interact with this cache

They don't depend on it.

> on purpose

Performance test is exactly this purpose. And since this PR is precisely a performance improvement, performance tests of course show it - I wouldn't call them "broken". They just need to be updated to test what were they supposed to test.