@iamamyth 2022-12-12
> But I want tests to run this code, this is how I found serialization issues (see conversation above). Maybe it's better to have something like rct::verRctClearCache() and call it where necessary?

You want some tests to run this code, and some (e.g. the crypto tests) to run the original. I find it best to reason from principles, not effects: Modifying the existing function changes its semantics (original: logically verify the transaction; new: retrieve verification info from cache or, on a miss, verify the semantics), which potentially breaks every call (the crypto-only tests demonstrate the breakage, but there's another obvious set of cases: users who either want a bigger cache or don't want one at all). You note the semantics mostly end up the same but that's actually the problem: load from cache and compute from first principles differ subtly, making for subtle, hard to find problems.