@SChernykh 2022-12-12
@UkoeHB I renamed the cached variant to `rct::verRctNonSemanticsSimpleCached`, all tests should be fine now.