@UkoeHB 2022-12-12
I'm still not excited about dumping this static cache thing into `rctSigs.cpp`. The cache is not extensible or reusable - for example notice that none of the seraphis code is in `rctSigs.cpp`.