@SChernykh 2022-12-12
I don't see how to generalize it without a lot of new boilerplate code. The code is just a few lines, and 50% of the new code is `get_hash` function which must be tailored for each new type. Yes, I can generalize it to something that caches any function of type `bool F(const T& data)`, but the size of code changes will double and `get_hash` will still stick out. I think that with the coming Seraphis, this whole code will need to be rewritten anyway, and this is when it should be generalized.