@UkoeHB 2022-12-12
I'm sorry to say it but this is how you get technical debt - one easy solution at a time.

> I think that with the coming Seraphis, this whole code will need to be rewritten anyway, and this is when it should be generalized.

The `ringct` directory is a dependency of `seraphis`, nothing related to signatures needs to be changed - you can check the [diffs](https://github.com/monero-project/monero/compare/master...UkoeHB:monero:seraphis_lib).

I don't know if you need to fully generalize and implement a template. You could write a tool that records {hash, verification result} and save the results in there without even telling the tool what kind of proof you validated. Then a little indirection to branch off `result_cache.has_result_for(tx_hash)` -> `..verify..; result_cache.add_record(tx_hash, result); return result;` or `return result_cache.get_result_for(tx_hash);`.