@kayabaNerve 2022-12-13
@UkoeHB That proposed solution isn't immediately implementable as TX validity will change on re-org. I just want to be clear about that again as it's easy to overlook.