@jeffro256 2023-01-11
With around 40 points of data for p2pool block propogations under real-world circumstances, running node and p2pool on same computer to minimize network delay randomness, this PR reduced block propogation round-trip time from 0.59 seconds to 0.12 seconds, a speedup of 4.9x.

Specs:
```
OS: Ubuntu 22
Storage: Internal SSD NVME
RAM: 32 GB
CPU: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
```