@luigi1111 2023-01-11
> @UkoeHB That proposed solution isn't immediately implementable as TX validity will change on re-org. I just want to be clear about that again as it's easy to overlook.

@kayabaNerve how does this PR handle reorgs?