@SChernykh 2023-01-11
@luigi1111 it's not affected by reorgs because it caches full tx data (key images and mixRings with full data, not indices).