@UkoeHB 2023-01-11
@luigi1111 it is only caching the result of verifying proofs. The result of checking key images and ring signatures against the chain is not cached.