@kayabaNerve 2023-01-12
This is currently fine so long as `do_serialize` serializes the generally un-serialized key images and mix ring from the RCT object, which is the assumption it relies on. My prior commentary was premised on if the ring isn't serialized, how the ring can change in a re-org and with it, we'd lose the context which made the proof valid.

As someone unsure the exact details of `do_serialize`, I would've written a test to confirm, yet so long as someone properly experienced confirms that property, I have no objections to this having been merged.