@jeffro256 2023-01-12
@kayabaNerve I don't know if you saw, but [this conversation](https://github.com/monero-project/monero/pull/8675#discussion_r1053636419) should cover this topic. As @vtnerd pointed out, the `do_serialize` function serializes *all* fields of `rctSig`, even the prunable stuff, which as you mentioned is not the "normal" serialization mode used for consensus. That function would be `rct::RctSigBase::serialize_rctsig_base`. I will write a unit test within the next couple days to confirm this behavior now that you mention it.